# docs/api/openapi.yaml
components:
  schemas:
    CatalogCreate:
      description: Schema for creating a new catalog.
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        identifier:
          maxLength: 255
          minLength: 1
          title: Identifier
          type: string
        themes:
          items:
            type: string
          title: Themes
          type: array
        title:
          maxLength: 255
          minLength: 1
          title: Title
          type: string
      required:
      - identifier
      - title
      title: CatalogCreate
      type: object
    CatalogListResponse:
      description: Schema for paginated catalog list responses.
      properties:
        items:
          items:
            $ref: '#/components/schemas/CatalogResponse'
          title: Items
          type: array
        limit:
          minimum: 1.0
          title: Limit
          type: integer
        offset:
          minimum: 0.0
          title: Offset
          type: integer
        total:
          minimum: 0.0
          title: Total
          type: integer
      required:
      - items
      - total
      - limit
      - offset
      title: CatalogListResponse
      type: object
    CatalogResponse:
      description: Schema for catalog responses.
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional description
          title: Description
        id:
          format: uuid
          title: Id
          type: string
        identifier:
          maxLength: 255
          minLength: 1
          title: Identifier
          type: string
        themes:
          description: List of theme identifiers
          items:
            type: string
          title: Themes
          type: array
        title:
          maxLength: 255
          minLength: 1
          title: Title
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
      required:
      - identifier
      - title
      - id
      - created_at
      - updated_at
      title: CatalogResponse
      type: object
    CatalogUpdate:
      description: Schema for updating an existing catalog.
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        themes:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Themes
        title:
          anyOf:
          - maxLength: 255
            minLength: 1
            type: string
          - type: 'null'
          title: Title
      title: CatalogUpdate
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    PurposeListResponse:
      description: Schema for paginated purpose list responses.
      properties:
        items:
          items:
            $ref: '#/components/schemas/PurposeResponse'
          title: Items
          type: array
        limit:
          minimum: 1.0
          title: Limit
          type: integer
        offset:
          minimum: 0.0
          title: Offset
          type: integer
        total:
          minimum: 0.0
          title: Total
          type: integer
      required:
      - items
      - total
      - limit
      - offset
      title: PurposeListResponse
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: X-API-Key
      type: apiKey
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact:
    email: mike@mikewaters.net
    name: LifeOS Development Team
    url: https://github.com/LifeOS/Substrate
  description: REST API for topic taxonomy and knowledge management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Ontology API
  version: 0.1.0
  x-api-id: ontology-api
  x-audience: internal
openapi: 3.1.0
paths:
...
  /catalogs:
    get:
      description: "List catalogs with pagination.\n\nArgs:\n    service: Catalog service dependency\n    limit: Maximum number\
        \ of results (1-100)\n    offset: Number of results to skip\n\nReturns:\n    Paginated list of catalogs"
      operationId: list_catalogs_catalogs_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogListResponse'
              example:
                items:
                - id: 00000000-0000-0000-378e-7dd8dd9e4e11
                  identifier: cat:tech-resources
                  title: Technology Resources
                  description: Collection of technical documentation, tutorials, and tools
                  themes:
                  - tech:software
                  - tech:ai
                  - tech:cloud
                  created_at: '2025-11-04T02:55:11.302186Z'
                  updated_at: '2025-11-04T02:55:11.302186Z'
                - id: 00000000-0000-0000-2b44-4d5e362b4a2a
                  identifier: cat:learning
                  title: Learning Materials
                  description: Educational content across various domains
                  themes:
                  - know:capture
                  - know:organize
                  - life:productivity
                  created_at: '2025-11-04T02:55:11.302186Z'
                  updated_at: '2025-11-04T02:55:11.302186Z'
                total: 2
                limit: 50
                offset: 0
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List catalogs
      tags:
      - catalogs
    post:
      description: "Create a new catalog.\n\nArgs:\n    catalog: Catalog creation data\n    service: Catalog service dependency\n\
        \nReturns:\n    Created catalog"
      operationId: create_catalog_catalogs_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a new catalog
      tags:
      - catalogs
  /catalogs/{catalog_id}:
    delete:
      description: "Delete a catalog.\n\nArgs:\n    catalog_id: Catalog ID\n    service: Catalog service dependency\n\nRaises:\n\
        \    NotFoundError: If catalog not found"
      operationId: delete_catalog_catalogs__catalog_id__delete
      parameters:
      - description: Catalog ID
        in: path
        name: catalog_id
        required: true
        schema:
          description: Catalog ID
          format: uuid
          title: Catalog Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete a catalog
      tags:
      - catalogs
    get:
      description: "Get a catalog by ID.\n\nArgs:\n    catalog_id: Catalog ID\n    service: Catalog service dependency\n\n\
        Returns:\n    Catalog\n\nRaises:\n    NotFoundError: If catalog not found"
      operationId: get_catalog_catalogs__catalog_id__get
      parameters:
      - description: Catalog ID
        in: path
        name: catalog_id
        required: true
        schema:
          description: Catalog ID
          format: uuid
          title: Catalog Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a catalog by ID
      tags:
      - catalogs
    put:
      description: "Update a catalog.\n\nArgs:\n    catalog_id: Catalog ID\n    updates: Fields to update\n    service: Catalog\
        \ service dependency\n\nReturns:\n    Updated catalog\n\nRaises:\n    NotFoundError: If catalog not found"
      operationId: update_catalog_catalogs__catalog_id__put
      parameters:
      - description: Catalog ID
        in: path
        name: catalog_id
        required: true
        schema:
          description: Catalog ID
          format: uuid
          title: Catalog Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update a catalog
      tags:
      - catalogs
...
  /purposes:
    get:
      description: "List purposes with pagination.\n\nArgs:\n    service: Purpose service dependency\n    limit: Maximum number\
        \ of results (1-100)\n    offset: Number of results to skip\n\nReturns:\n    Paginated list of purposes"
      operationId: list_purposes_purposes_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurposeListResponse'
              example:
                items:
                - id: 00000000-0000-0000-a68c-d152bbfe92d2
                  identifier: purpose:reference
                  title: Reference Material
                  description: Material kept for future reference and lookup
                  role: reference
                  meaning: To be consulted when needed
                  created_at: '2025-11-04T02:55:11.302186Z'
                  updated_at: '2025-11-04T02:55:11.302186Z'
                total: 1
                limit: 50
                offset: 0
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List purposes
      tags:
      - purposes
    post:
      description: "Create a new purpose.\n\nArgs:\n    purpose: Purpose creation data\n    service: Purpose service dependency\n\
        \nReturns:\n    Created purpose"
      operationId: create_purpose_purposes_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurposeCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurposeResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a new purpose
      tags:
      - purposes
...
servers:
- description: Development server
  url: http://localhost:8000
- description: Production server
  url: https://api.lifeos.dev